utils/requirements.txt
pytesseract>=0.2.6
Pillow>=5.4.1
pdf2image>=1.5.4